+++ CHANGELOG.md
@@ -4,13 +4,20 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [4.5.2] - Unreleased
+## [4.5.2] - 2023-05-18
 
 ### Added
 
 - Missing staves for all Medenia classes (AB 50+)
 - Missing staves for bards (AB 70+)
 - Missing staves for summoners (AB 70+)
+
+### Fixed
+
+- Cooldown detection fixed on 64-bit computers
+- Macro toolbar state now updates more reliably (map/location change)
+- Spell queue now hides when last character logs out
+- Non-integer window scaling now supported (ex: 150%, 175%, etc)
 
 ## [4.5.1] - 2023-04-18